#CLOUD-138 fixed assert while stoppping async tcp server with existing connections
#CLOUD-141 cloud_db now adds json message body to 401 unauthorized reply
#CLOUD-142 main processors wait for async call completion when stopping process
#CLOUD-144 authentication failure due to account not activated results in accountNotActivated result code
#CLOUD-145 system can be unbinded from cloud account using system credentials
#CLOUD-146 added /account/reactivate request to re-send activation link for existing account
#CLOUD-146 forbidden reactivation of activated account
#CLOUD-73 implemented /account/reset_password request
#CLOUD-73 password reset code usage is restricted to /account/update only
#CLOUD-73 temporary account password logic moved to separate class TemporaryAccountPasswordManager. temporary password are now removed from DB on expiration
AsyncOperationGuard with test
Merge with dev_3.0
Merge with dev_3.0_pleaseStop
SocketGlobals pleaseStop refactoring
branch
class QnSemaphore moved to nx_utils library
cloud_db returns json with resultCode in 'success' response
fixed compile error
fixed compile warning
fixed error in test
fixed linux build
fixed msvc compile errors
fixed release build
fixed some compile warnings
fixed: reset_password notification was not sent
merged heads
merged into dev_3.0
minor cleanup
minor fixes
more tests
noError result code replaced with ok
push it
release compile fixes
stun attributes rebranding
stun::AsyncClient connection interface refactoring
true windows compile fix
windows compile fixes

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ unit_tests/common_ut/src/utils/network/http/multipart_content_parser_ut.cpp
@@ -196,7 +196,7 @@ TEST( HttpMultipartContentParser, unSizedDataSimple )
             std::make_shared<CustomOutputStream<decltype(decodedFramesProcessor)> >(
                 decodedFramesProcessor ) );
 
-        for( size_t pos = 0; pos < testData.size(); pos += dataStep )
+        for( int pos = 0; pos < testData.size(); pos += dataStep )
             parser.processData( QnByteArrayConstRef(
                 testData,
                 pos,
@@ -265,7 +265,7 @@ TEST( HttpMultipartContentParser, unSizedData )
                 std::make_shared<CustomOutputStream<decltype(decodedFramesProcessor)> >(
                     decodedFramesProcessor ) );
 
-            for( int pos = 0; pos < testData.size(); pos += dataStep )
+            for( size_t pos = 0; pos < (size_t)testData.size(); pos += dataStep )
                 parser.processData( QnByteArrayConstRef(
                     testData,
                     pos,
